+++ prequ/scripts/prequ.py
@@ -1,7 +1,7 @@
 from .. import click
 from . import build_wheels
 from . import compile_in
-from . import compile_all
+from . import compile
 from . import sync
 from . import update
 
@@ -14,7 +14,7 @@ def main():
 
 main.add_command(build_wheels.main, 'build-wheels')
 main.add_command(compile_in.cli, 'compile-in')
-main.add_command(compile_all.main, 'compile-all')
+main.add_command(compile.main, 'compile')
 main.add_command(sync.cli, 'sync')
 main.add_command(update.main, 'update')